move websites pages to website namespace
