--- app/controllers/static_pages_controller.rb
@@ -1,17 +1,0 @@
-class StaticPagesController < ApplicationController
-  skip_before_action :authenticate_agent!
-
-  layout "website"
-
-  def welcome
-    redirect_to(organisations_path) if current_agent
-  end
-
-  def legal_notice; end
-
-  def cgu; end
-
-  def privacy_policy; end
-
-  def accessibility; end
-end
