+++ app/controllers/website/base_controller.rb
@@ -1,0 +1,5 @@
+module Website
+  class BaseController < ApplicationController
+    layout "website"
+  end
+end
