+++ app/views/layouts/application.html.erb
@@ -1,5 +1,5 @@
 <% content_for :header do %>
-  <%= render 'common/header', is_within_application: true %>
+  <%= render 'common/header' %>
 <% end %>
 
 <%= render "layouts/application_base" do %>
