+++ app/views/layouts/website.html.erb
@@ -1,5 +1,5 @@
 <% content_for :header do %>
-  <%= render 'common/header', is_within_application: false %>
+  <%= render 'common/header_website' %>
 <% end %>
 
 <%= render "layouts/application_base" do %>
