+++ app/views/website/static_pages/cgu.html.erb
