+++ app/views/website/stats/_header.html.erb
