+++ app/views/website/stats/_stats.html.erb
