+++ app/views/website/stats/deployment_map.html.erb
