+++ app/views/website/stats/index.html.erb
