+++ spec/controllers/website/static_pages_controller_spec.rb
@@ -1,4 +1,4 @@
-describe StaticPagesController do
+describe Website::StaticPagesController do
   render_views
 
   let!(:organisation) { create(:organisation) }
